+++ crates/jsonschema/src/keywords/const_.rs
@@ -49,19 +49,6 @@ impl Validate for ConstArrayValidator {
         }
     }
 }
-impl core::fmt::Display for ConstArrayValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "const: [{}]",
-            self.value
-                .iter()
-                .map(ToString::to_string)
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-}
 
 struct ConstBooleanValidator {
     value: bool,
@@ -101,11 +88,6 @@ impl Validate for ConstBooleanValidator {
         }
     }
 }
-impl core::fmt::Display for ConstBooleanValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "const: {}", self.value)
-    }
-}
 
 struct ConstNullValidator {
     schema_path: JSONPointer,
@@ -137,11 +119,6 @@ impl Validate for ConstNullValidator {
     #[inline]
     fn is_valid(&self, instance: &Value) -> bool {
         instance.is_null()
-    }
-}
-impl core::fmt::Display for ConstNullValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "const: {}", Value::Null)
     }
 }
 
@@ -192,12 +169,6 @@ impl Validate for ConstNumberValidator {
     }
 }
 
-impl core::fmt::Display for ConstNumberValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "const: {}", self.original_value)
-    }
-}
-
 pub(crate) struct ConstObjectValidator {
     value: Map<String, Value>,
     schema_path: JSONPointer,
@@ -243,20 +214,6 @@ impl Validate for ConstObjectValidator {
     }
 }
 
-impl core::fmt::Display for ConstObjectValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "const: {{{}}}",
-            self.value
-                .iter()
-                .map(|(key, value)| format!(r#""{}":{}"#, key, value))
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-}
-
 pub(crate) struct ConstStringValidator {
     value: String,
     schema_path: JSONPointer,
@@ -296,12 +253,6 @@ impl Validate for ConstStringValidator {
         } else {
             false
         }
-    }
-}
-
-impl core::fmt::Display for ConstStringValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "const: {}", self.value)
     }
 }