+++ crates/jsonschema/src/keywords/dependencies.rs
@@ -5,7 +5,7 @@ use crate::{
     paths::{JSONPointer, JsonPointerNode},
     primitive_type::PrimitiveType,
     schema_node::SchemaNode,
-    validator::{format_key_value_validators, Validate},
+    validator::Validate,
 };
 use serde_json::{Map, Value};
 
@@ -79,16 +79,6 @@ impl Validate for DependenciesValidator {
         } else {
             no_error()
         }
-    }
-}
-
-impl core::fmt::Display for DependenciesValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "dependencies: {{{}}}",
-            format_key_value_validators(&self.dependencies)
-        )
     }
 }
 
@@ -173,15 +163,6 @@ impl Validate for DependentRequiredValidator {
         }
     }
 }
-impl core::fmt::Display for DependentRequiredValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "dependentRequired: {{{}}}",
-            format_key_value_validators(&self.dependencies)
-        )
-    }
-}
 
 pub(crate) struct DependentSchemasValidator {
     dependencies: Vec<(String, SchemaNode)>,
@@ -241,15 +222,6 @@ impl Validate for DependentSchemasValidator {
     }
 }
 
-impl core::fmt::Display for DependentSchemasValidator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "dependentSchemas: {{{}}}",
-            format_key_value_validators(&self.dependencies)
-        )
-    }
-}
 #[inline]
 pub(crate) fn compile<'a>(
     _: &'a Map<String, Value>,